refactorize code in main.go
